+++ services/base.py
@@ -3,7 +3,7 @@ from typing import Protocol
 
 class LLMService(Protocol):
     model: str | None
-    latency_ms: float | None
+    latency_ms: int | None
 
     async def generate_caption(
         self, prompt: str, system_instruction: str
